requirements.txt
fastapi
tensorflow
pillow
numpy
uvicorn
